🐳 Disabled firebase CLI installation temporarily
I need to advance with stuff and firebase CLI fails to install.
If I use pnpm, it says `pnpm setup` should be run. I have to run it with `RUN SHELL=/bin/bash ...`, otherwise it says it doesn't recognize the shell type.
If I do that, it runs setup successfully, but then, using pnpm to install firebase fails because it doesn't recognize the setup which was just executed for some reason.
When installing with npm, it installs but takes a long time. Also, the `firebase` command isn't available from the terminal (WTF?? 🤦)